Adjust drag inertia frame start always to the same point at bottom of screen
The point is arbitrary chosen at the bottom of screen. This doesn't impact behavior when tilt is low and helps with large tilt values.
The important part is not to change starting point (previousTouchLocation) at every frame to points above as they increase map velocity.

decelerationFactor is very large (doesn't feel like it is correctly used in our code compared to platform scroll views). So, adjusted hereto 0.97 (original decelerationFactor is 0.998).